// src/services/game/game.interface.ts
export interface IGame {
 avatarInfo: string;
 info: string;
 price: string;
 category: string;
 about: string;
 id: string;
 description: string;
 title: string;
 genre: string;
 developer: string;
 puslisher: string;
 franchise: string;
 releaseDate: string;
 images: string[];
 imageHeader: string;
}